add logout component
